# docs-site/public/admin-api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IdentityCreate": {
        "properties": {
          "agent_type": {
            "default": "custom",
            "title": "Agent Type",
            "type": "string"
          },
          "api_key": {
            "title": "Api Key",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "mapped_pg_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mapped Pg Role"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "team": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team"
          }
        },
        "required": [
          "name",
          "api_key"
        ],
        "title": "IdentityCreate",
        "type": "object"
      },
      "IdentityResponse": {
        "description": "Response model - deliberately excludes api_key_hash.",
        "properties": {
          "agent_type": {
            "title": "Agent Type",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "mapped_pg_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mapped Pg Role"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "team": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "name",
          "agent_type",
          "roles",
          "enabled",
          "metadata"
        ],
        "title": "IdentityResponse",
        "type": "object"
      },
...
      "SourceRoleGrantCreate": {
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "role_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Id"
          },
          "role_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Key"
          },
          "source_id": {
            "minLength": 1,
            "title": "Source Id",
            "type": "string"
          }
        },
        "required": [
          "source_id"
        ],
        "title": "SourceRoleGrantCreate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Management plane for InterLock",
    "title": "InterLock Admin",
    "version": "1.0.0rc12"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/identities": {
      "get": {
        "operationId": "list_identities_api_identities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IdentityResponse"
                  },
                  "title": "Response List Identities Api Identities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Identities",
        "tags": [
          "identities"
        ]
      },
      "post": {
        "operationId": "create_identity_api_identities_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Identity",
        "tags": [
          "identities"
        ]
      }
    },
    "/api/identities/{identity_id}": {
      "delete": {
        "operationId": "delete_identity_api_identities__identity_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "identity_id",
            "required": true,
            "schema": {
              "title": "Identity Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Identity",
        "tags": [
          "identities"
        ]
      },
      "get": {
        "operationId": "get_identity_api_identities__identity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "identity_id",
            "required": true,
            "schema": {
              "title": "Identity Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Identity",
        "tags": [
          "identities"
        ]
      },
      "put": {
        "operationId": "update_identity_api_identities__identity_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "identity_id",
            "required": true,
            "schema": {
              "title": "Identity Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Identity",
        "tags": [
          "identities"
        ]
      }
    },
    "/api/identities/{identity_id}/source-role-grants": {
      "get": {
        "description": "What this agent can reach. Revoked grants are history, not access, so\nthey are excluded unless asked for.",
        "operationId": "list_identity_grants_api_identities__identity_id__source_role_grants_get",
        "parameters": [
          {
            "in": "path",
            "name": "identity_id",
            "required": true,
            "schema": {
              "title": "Identity Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include_revoked",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Revoked",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceRoleGrantResponse"
                  },
                  "title": "Response List Identity Grants Api Identities  Identity Id  Source Role Grants Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Identity Grants",
        "tags": [
          "identities"
        ]
      },
      "post": {
        "description": "Grant a role on a source, without touching the identity's API key.\n\nIdempotent: granting a role the identity already holds restates its terms\nand re-enables it if it had been revoked.",
        "operationId": "create_identity_grant_api_identities__identity_id__source_role_grants_post",
        "parameters": [
          {
            "in": "path",
            "name": "identity_id",
            "required": true,
            "schema": {
              "title": "Identity Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceRoleGrantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRoleGrantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Identity Grant",
        "tags": [
          "identities"
        ]
      }
    },
    "/api/identities/{identity_id}/source-role-grants/{grant_id}": {
      "delete": {
        "description": "Withdraw a grant. The row is kept, disabled.\n\nDeleting it would make \"did this agent ever hold access, and when was it\nwithdrawn\" unanswerable, which is the question an incident review asks.\nThe agent is not re-keyed and its other grants are untouched; the change\ntakes effect on its next request.",
        "operationId": "revoke_identity_grant_api_identities__identity_id__source_role_grants__grant_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "identity_id",
            "required": true,
            "schema": {
              "title": "Identity Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "grant_id",
            "required": true,
            "schema": {
              "title": "Grant Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Revoke Identity Grant Api Identities  Identity Id  Source Role Grants  Grant Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Identity Grant",
        "tags": [
          "identities"
        ]
      }
    },
...
    "/dashboard/access-control/identities/{ident_id}": {
      "get": {
        "operationId": "identity_detail_dashboard_access_control_identities__ident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ident_id",
            "required": true,
            "schema": {
              "title": "Ident Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Identity Detail",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/dashboard/discovery/entities/{entity_type}/{entity_text}": {
      "get": {
        "description": "Entity detail: assets that reference it, co-occurring entities.",
        "operationId": "discovery_entity_detail_dashboard_discovery_entities__entity_type___entity_text__get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_type",
            "required": true,
            "schema": {
              "title": "Entity Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entity_text",
            "required": true,
            "schema": {
              "title": "Entity Text",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discovery Entity Detail",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/dashboard/entities": {
      "get": {
        "description": "Entity explorer page. Shows all entities by default, filtered on search.",
        "operationId": "entities_page_dashboard_entities_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Entities Page",
        "tags": [
          "dashboard"
        ]
      }
    },
...
  }
}